codequality cont

+++ Project/config_folder/globals/globals.go
@@ -2,10 +2,10 @@ package globals
 
 // ------------------ CONST ------------------
 const (
-	NumElevators     int     = 3
-	NumFloors        int     = 4
-	NumButtonTypes   int     = 3
-	DoorOpenDuration float64 = 4.0 // [s] open door duration
+	NumElevators     int = 3
+	NumFloors        int = 4
+	NumButtonTypes   int = 3
+	DoorOpenDuration int = 4 // [s] open door duration
 
 	BackupFile string = "SystemBackup.txt"
 	BackupDir  string = "BackupFiles"
